requirements.txt
# Web 框架
fastapi>=0.111.0
uvicorn[standard]>=0.30.0

# ORM / 数据库
sqlalchemy>=2.0.30
pydantic>=2.8.2
pydantic-settings>=2.4.0
psycopg[binary]>=3.1        # ← v3 驱动，匹配 postgresql+psycopg
asyncpg>=0.29               # ← 异步引擎驱动（tests/quick 使用）

# Migrations
alembic>=1.13.0

# 密码哈希
passlib>=1.7.4
bcrypt>=4.1.2
argon2-cffi>=23.1.0

# JWT
python-jose[cryptography]>=3.3.0
cryptography>=42.0.0

# 类型检查
mypy>=1.10.0
types-setuptools

# 代码质量工具
pre-commit>=3.7.0
ruff>=0.5.0
black>=24.4.0
isort>=5.13.0

# 其他
PyYAML>=6.0.1
email-validator
httpx>=0.27
celery[redis]
prometheus-client>=0.20.0

# --- Phase 2.8: OpenTelemetry deps ---
opentelemetry-api~=1.27.0
opentelemetry-sdk~=1.27.0
opentelemetry-exporter-otlp-proto-grpc~=1.27.0
opentelemetry-exporter-otlp-proto-http~=1.27.0   # 新增：支持 OTLP/HTTP → Jaeger:4318

# instrumentations（与 SDK 版本匹配的 beta 包）
opentelemetry-instrumentation-fastapi~=0.48b0
opentelemetry-instrumentation-sqlalchemy~=0.48b0
opentelemetry-instrumentation-redis~=0.48b0
opentelemetry-instrumentation-celery~=0.48b0

# --- Test Stack（本地与 CI 统一） ---
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=7.0.0
pytest-env>=1.2.0    # 让 pytest 识别 pytest.ini/pyproject 的 [env] 注入
